Basic User Password Authentication
1) Included changes to support multiple roles
2) Basic username and password authentication
3) New authentication Controller

+++ src/main/java/com/dyvr/librarysystem/Library/System/service/BookService.java
@@ -27,9 +27,13 @@ public class BookService {
        return bookRepository.save(book);
     }
 
-    public String deleteBookById(int id){
-        bookRepository.deleteById(id);
-        return "Book with id {" + id + "} removed";
+    public Boolean deleteBookById(int id){
+        Optional<Book> book = bookRepository.findById(id);
+        if (book.isPresent()) {
+            bookRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
 }
